Add VS 2012 and VS 2013 versions of libprotobuf.lib
Adding these libraries lets the source-sdk be built with VS 2012 and VS
2013, in addition to VS 2010. Just copy createallprojects.bat or
creategameprojects.bat and add /2012 or /2013 to the vpc command line.
Then load the resultant .sln file into the appropriate version of Visual
Studio. A full rebuild may be necessary the first time in order to clear
out .obj and .pdb files from previous builds.